pyproject.toml
[tool.poetry]
name = "ragn"
version = "0.1.0"
description = ""
authors = ["Caio Dadauto <caiodadauto@gmail.com>"]

[tool.poetry.dependencies]
python = ">3.9"
graph-nets = {git = "https://github.com/caiodadauto/graph_nets.git", rev = "feature/recurrent_blocks"}
tensorflow = "^2.16.1"
dm-sonnet = "^2.0.2"
tensorflow-probability = "^0.24.0"
matplotlib = "^3.8.4"
tqdm = "^4.66.2"
numba = "^0.59.1"
scikit-learn = "^1.4.2"
joblib = "^1.4.0"
mlflow = "^2.12.2"
hydra-core = "^1.3.2"
seaborn = "^0.13.2"

[tool.poetry.dev-dependencies]
pytest = "^5.2"

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
